Re-organised GUI folder and fixed imports (ignore review)
