+++ miezepy/gui/py_gui/gui_common/python_syntax.py
